Rename property.py → fancy_property.py
TIL: It is not a good thing to use a module having the same name as a
     commonly used decorator.
