+++ sipa/model/default/__init__.py
@@ -4,7 +4,7 @@
 from abc import ABCMeta, abstractmethod
 from collections import namedtuple
 
-from sipa.model.property import active_prop
+from sipa.model.fancy_property import active_prop
 
 
 # noinspection PyMethodMayBeStatic
