+++ sipa/model/fancy_property.py
