+++ sipa/model/gerok/user.py
@@ -8,7 +8,7 @@ from flask.globals import current_app
 from werkzeug.local import LocalProxy
 
 from sipa.model.default import BaseUser
-from sipa.model.property import active_prop, connection_dependent, \
+from sipa.model.fancy_property import active_prop, connection_dependent, \
     unsupported_prop
 from sipa.utils import argstr
 from sipa.utils.exceptions import PasswordInvalid, UserNotFound
